+++ protected/humhub/modules/marketplace/widgets/ModuleUpdateActionButtons.php
@@ -44,6 +44,7 @@ class ModuleUpdateActionButtons extends Widget
 
         $html .= Button::light(Yii::t('MarketplaceModule.base', 'Update'))
             ->link(['/marketplace/update/install', 'moduleId' => $this->module->id])
+            ->cssClass('link-accent')
             ->sm()
             ->action('marketplace.update');