+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 HumHub Changelog
 ================
 
-1.18.0-beta.3 (TBD)
--------------------
+1.18.0-beta.3 (September 9, 2025)
+---------------------------------
 
 > Updated minimum required PHP version to 8.2.
 
@@ -13,6 +13,7 @@ HumHub Changelog
 - Fix #7680: Migration of Theme Colors broken
 - Fix #7681: Add missing 'DatabaseForm' to 'InstallController
 - Fix #7685: In custom themes, the default colors in the Appearance Settings are overwritten by the base theme colors
+- Chg: Link color is now black. Use the `link-accent` class to emphasize a link.
 - Fix #7688: Fix notifications filter toggler
 
 1.18.0-beta.2 (September 4, 2025)